% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server_helper.R
\name{check_completeness}
\alias{check_completeness}
\title{check completeness}
\usage{
check_completeness(candidate_submission = NULL, registration_questions = NULL)
}
\description{
check if all attributes are complete and valid
}
\author{
Farid Azouaou
}
