% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server_helper.R
\name{finalize_submission}
\alias{finalize_submission}
\title{finalize training booking submission}
\usage{
finalize_submission(registration_questions = NULL, candidate_submission = NULL)
}
\description{
finalize training booking submission
}
\author{
Farid Azouaou
}
