% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server_helper.R
\name{import_data_mongo_db}
\alias{import_data_mongo_db}
\title{import_data_mongo_db}
\usage{
import_data_mongo_db(
  base_url = NULL,
  mongo_db_api_key = NULL,
  target_db = "DB_Full_byline",
  target_table = NULL,
  target_filter = NULL,
  req_limit = 4000
)
}
\arguments{
\item{base_url}{a base url for mongoDB DATA API}
}
\description{
import data from mongo DB databse stored on cloud based on some criteria
}
