% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server_helper.R
\name{upload_to_database}
\alias{upload_to_database}
\title{upload_to_database}
\usage{
upload_to_database(
  input_data = NULL,
  upload_mode = "overwrite",
  sc_meta_data = NULL
)
}
\arguments{
\item{input_data}{input_data}
}
\description{
t.b.d
}
